+++ client/web/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./reset.css";
-import { QueryProvider } from "../trpc/provider";
+import { QueryProvider } from "../common/trpc/provider";
 
 const inter = Inter({ subsets: ["latin"] });
