+++ client/web/src/common/trpc/provider.tsx
@@ -14,7 +14,7 @@ export const QueryProvider = ({ children }: Props) => {
     trpcReact.createClient({
       links: [
         httpBatchLink({
-          url: "http://localhost:3000/trpc",
+          url: "/trpc",
           // You can pass any HTTP headers you wish here
           async headers() {
             return {};
